+++ Sources/BroadCore/Domain/Logging/BroadLoggerProtocol.swift
@@ -1,3 +1,0 @@
-public protocol BroadLoggerProtocol: Sendable {
-    func log(_ event: BroadLogEvent)
-}