+++ Sources/BroadCore/Domain/Models/BroadCoreModule.swift
@@ -1,7 +1,0 @@
-public struct BroadCoreModule: Equatable, Sendable {
-    public let identifier: String
-
-    public init(identifier: String = "BroadCore") {
-        self.identifier = identifier
-    }
-}