+++ Sources/BroadCore/Domain/UseCases/RequestTrackingAuthorizationUseCaseProtocol.swift
@@ -1,5 +1,0 @@
-public protocol TrackingAuthorizationUseCaseProtocol: Sendable {
-    @MainActor
-    @discardableResult
-    func callAsFunction() async -> TrackingAuthorizationStatus
-}