+++ Sources/BroadCore/Infrastructure/Logging/NoOpBroadLogger.swift
@@ -1,5 +1,0 @@
-public struct NoOpBroadLogger: BroadLoggerProtocol {
-    public init() {}
-
-    public func log(_: BroadLogEvent) {}
-}